5 test room logic

+++ src/main/java/com/example/yacht_backend/service/RoomService.java
@@ -4,6 +4,8 @@ import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.web.context.request.async.DeferredResult;
 import com.example.yacht_backend.model.ActiveRoom;
@@ -15,46 +17,56 @@ import com.example.yacht_backend.domain.RoomData;
 
 @Service
 public class RoomService {
+    private static final Logger logger = LoggerFactory.getLogger(RoomService.class);
     private final RoomDatabaseService roomDatabaseService;
     private final UserDatabaseService userDatabaseService;
-    private final ConcurrentHashMap<String, RoomData> roomGuestMap = new ConcurrentHashMap<>();
+    private final ConcurrentHashMap<String, RoomData> roomGuestMap;
+    private Long deferredResultTimeout = 60000L;
+    private boolean onCompletion = true;
 
-    RoomService(RoomDatabaseService roomDatabaseService, UserDatabaseService userDatabaseService) {
+    RoomService(RoomDatabaseService roomDatabaseService, UserDatabaseService userDatabaseService, ConcurrentHashMap<String, RoomData> roomGuestMap) {
         this.roomDatabaseService = roomDatabaseService;
         this.userDatabaseService = userDatabaseService;
+        this.roomGuestMap = roomGuestMap;
+    }
+
+    public void setDefferedResultTimeout(Long deferredResultTimeout) {
+        this.deferredResultTimeout = deferredResultTimeout;
+    }
+
+    public void enableOnCompletion(boolean onCompletion) {
+        this.onCompletion = onCompletion;
     }
 
     public List<PendingRoom> getAllRooms() {
+        logger.info("RoomService.getAllRooms 실행");
         return roomDatabaseService.findAllPendingRoom();
     }
     
-    public CreateNewRoomResponse createNewRoom(String sessionId, String hostData) {
+    public DeferredResult<CreateNewRoomResponse> createNewRoom(String sessionId, String hostData) {
+        logger.info("RoomService.createNewRoom 실행");
         String userId = userDatabaseService.findUserIdBySessionId(sessionId);
-        ActiveRoom guestRoom = roomDatabaseService.findRoomByGuestUserId(userId);
+        ActiveRoom guestRoom = roomDatabaseService.findActiveRoomByGuestUserId(userId);
         
         if (guestRoom != null) {
-            DeferredResult<String> guestId = new DeferredResult<String>(60000L);
-            guestId.setResult(null);
-            DeferredResult<String> data = new DeferredResult<String>(60000L);
-            data.setResult(null);
-            return new CreateNewRoomResponse(null, guestId, data);
+            DeferredResult<CreateNewRoomResponse> createNewRoomResponse = new DeferredResult<>();
+            createNewRoomResponse.setResult(new CreateNewRoomResponse(null, null, null));
+            return createNewRoomResponse;
         }
         String roomId = UUID.randomUUID().toString(); // 임시로 임의의 룸을 생성
-        DeferredResult<String> guestUserId = new DeferredResult<String>(60000L);
-        DeferredResult<String> data = new DeferredResult<String>();
+        DeferredResult<CreateNewRoomResponse> createNewRoomResponse = new DeferredResult<>(deferredResultTimeout);
         
-        RoomData roomData = new RoomData(roomId, userId, hostData, guestUserId, data);
+        RoomData roomData = new RoomData(roomId, userId, hostData, createNewRoomResponse);
         roomGuestMap.put(roomId, roomData);
         PendingRoom pendingRoom = new PendingRoom(roomId, userId, hostData);
         roomDatabaseService.save(pendingRoom);
 
         // timeout인 경우
-        guestUserId.onTimeout(() -> {
+        createNewRoomResponse.onTimeout(() -> {
             synchronized (roomData) {
                 if(roomData.isOpen()) {
                     roomData.close();
-                    guestUserId.setResult(null);
-                    data.setResult(null);
+                    createNewRoomResponse.setResult(new CreateNewRoomResponse(null, null, null));
                 }
                 roomGuestMap.remove(roomId);
                 roomDatabaseService.delete(pendingRoom);
@@ -62,26 +74,29 @@ public class RoomService {
         });
 
         // 어떤 이유에서든(네트워크 에러, 클라이언트쪽 timeout 등) 연결이 끊긴 경우 or 처리가 완료된 경우
-        guestUserId.onCompletion(() -> {
-            synchronized (roomData) {
-                if (roomData.isOpen()) {
-                    roomData.close();
-                    if (guestUserId.hasResult()) {
-                        roomDatabaseService.save(new ActiveRoom(roomId, userId, (String)guestUserId.getResult()));
+        createNewRoomResponse.onCompletion(() -> {
+            if (onCompletion) {
+                synchronized (roomData) {
+                    if (roomData.isOpen()) {
+                        roomData.close();
+                        if (createNewRoomResponse.hasResult()) {
+                            CreateNewRoomResponse response = (CreateNewRoomResponse)createNewRoomResponse.getResult();
+                            roomDatabaseService.save(new ActiveRoom(roomId, userId, response.getGuestUserId()));
+                        }
+                        else {
+                            createNewRoomResponse.setResult(new CreateNewRoomResponse(null, null, null));
+                        }
                     }
-                    else {
-                        guestUserId.setResult(null);
-                        data.setResult(null);
-                    }
+                    roomGuestMap.remove(roomId);
+                    roomDatabaseService.delete(pendingRoom);
                 }
-                roomGuestMap.remove(roomId);
-                roomDatabaseService.delete(pendingRoom);
             }
         });
-        return new CreateNewRoomResponse(roomId, guestUserId, data);
+        return createNewRoomResponse;
     }
 
     public EnterRoomResponse enterRoom(String roomId, String sessionId, String guestData) {
+        logger.info("RoomService.enterRoom 실행");
         String userId = userDatabaseService.findUserIdBySessionId(sessionId);
         if(roomDatabaseService.isUserInRoom(userId)) {
             return new EnterRoomResponse("Already in room", false);
@@ -95,14 +110,12 @@ public class RoomService {
         synchronized (roomData) {
             if (roomData.isOpen()) {
                 roomData.close();
-                String hostUserId = roomData.getHostUserData();
-                DeferredResult<String> guestUserId = roomData.getGuestUserId();
-                DeferredResult<String> guestUserData = roomData.getGuestUserData();
-                if (guestUserId.hasResult()) {
+                String hostUserId = roomData.getHostUserId();
+                DeferredResult<CreateNewRoomResponse> createNewRoomResponse = roomData.getCreateNewRoomResponse();
+                if (createNewRoomResponse.hasResult()) {
                     return new EnterRoomResponse("Room is full", false);
                 }
-                guestUserId.setResult(userId);
-                guestUserData.setResult(guestData);
+                createNewRoomResponse.setResult(new CreateNewRoomResponse(roomId, userId, guestData));
                 
                 return new EnterRoomResponse(hostUserId, true);
             }
